Renamed run-pc_local.csh to run-pc_mpi.csh to better reflect the difference between run-pc.csh (generic, simple CAM 'run' script) and this version.
run-pc_mpi explicitly sets environment variables to be sure they're set on
each processor (NFS) (for bangkok /scratch/local).  A single CAM is run on a
node using both processors on that node. Also, this script is NOT for use with
central CAM directory.

This script will also attempt to build cam if the executable does not exist.
Not for the faint of heart.


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@1598 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 2625f7eff267941f4a23d6d98ae5d978fe91e9f1