Major commit of a variety of updates.   The details are spelled out in the top level README file, and pointers to the major changes are listed on the main DART html page.  The brief version is:
1. The preprocess program for generating observation specific files has
been improved and simplified.  See the README, and the html files in the
obs_def and obs_kind directories for more details.  It should be simpler
to define new obs_types, and you no longer need to edit path_names files
as you add or delete obs types.

2. Observation sequence files now write out a table of contents for only 
the observation types which occur in that file, and set 'max_obs' to the
actual number of observations in the file.

3. The MPI parallel version of filter should now kill other tasks if one
exits.  This may mean adding null_mpi_utilities_mod.f90 to any private
path_names files you have; all the ones in the repository have been updated.

4. The mpi_utilities modules should now compile on all systems as they are
distributed, without editing, except for users of the gfortran compiler.
There is a shell program 'fixsystem' to edit the files for gfortran use.

5. All model work directories now have a 'quickbuild.csh' script which
recompiles all executables in that directory.  The 'workshop_setup.csh'
scripts have been removed from the models which are not used by the
tutorial workshop.  In the future they may become more customized for the
tutorials, and so for general use the quickbuild script is recommended.

6. Some of the large model directories now have a 'full_experiment'
directory where the more complicated scripts are placed for multi-day
filter advances.  The single filter run scripts remain in the previous
'shell_scripts' directory for each model.

7. The restart_file_utility now prints out times in both gregorian day count
format and the more recognizable Y/M/D format.

8. We have made an attempt to minimize the amount of spurious output that
goes into the dart_log.out file, so it is easier to see real error messages
and important status information.  Namelist output defaults to a file
'dart_log.nml' and you can turn off the module registration messages via
the utilities_nml namelist.  The location box statistics default to off.

9. The names of the state vector files for model advances now match the
convention for restart files when there is more than one file.
This change should be invisible to users, and increases the usefulness 
of the restart_file_utility program.

10. The test_dart.csh script has been updated.  If the MPI shell environment
variable is defined, the bgrid_solo model will be compiled with MPI.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@3414 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 3172fe897d7fb44847a6648ef75902ab2d66e2e9