Fixed errors in the obs_sequence_tool, the ncep observation converters, and obs_diag when restricting locations to a particular bounding box.  The case was unlikely (a box which started west of 0 and crossed east past 180) but the previous code would have done an incorrect selection of locations.
A single common routine was added to the utilities module which 
handles the wrap across longitude 0 consistently, and code in the
locations mod now calls it.  Other converters and code which
deal with longitude wrap have been updated to use either the
utilities routine (which deals with simple values) or the 
locations mod routine (which takes location_type arguments).


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@3809 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 464d705cc58caa2338f6cb1c4054a83b346e6f9a