Updates to the MADIS converters:  whether or not the QC values are in the files is detected at runtime.  There is a internal variable which can be set at compile time to force the programs to ignore QC fields even if present and use all observations, but by default the QC will be used if it exists.
The single surface converter has been split into two versions;
one which reads the standard METAR files and one which reads
the very similar but slightly different Mesonet files.

Add the additional METAR_ALTIMETER and METAR_ moisture
types.  For now, THERE ARE NO MESONET_xxx types -- i still haven't
gotten a clear consensus on whether this was desired or not.  
For now, this version of the MESONET converters will use LAND_SFC_ 
as the output type for mesonet files.  


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@4282 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: c99078ee8288711dfe19dd242ab9cb5b94623635