This advances the LANL/POP 2_0_1 version of POP ... but still cannot be used for assimilation until the POP code is modified to do a forward euler timestep.  This compiles and runs on coral (SLES10) with the ifort (IFORT) 10.1 20090203 compiler with the following flags: FFLAGS = -O0 -fpe0 -vec-report0 -assume byterecl
Coral is an Intel-based machine - so all binary files were little-endian.
I like to append a ".le" suffix on those files.

It was checked in the gx3v5 configuration.

For coral with the openmpi framework, it is necessary to
specify input.nml:&mpi_utilities_nml:reverse_task_layout = .true., 
For bluefire ... it must be .false. (the default).



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@4398 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 34b9b93e11947a9b363c1bcb9d0d6a9835c985b8