fix a bug in the termination code - if the end time was exactly one window before 0Z, it would fail the end test and loop forever.
also add an option (off by default for backwards compatibility)
to allow the script to loop over missing input files.  if some
time periods are missing some types of input files there is now
an option to complain on standard output but continue to run.

also added a couple tests to be sure the required executables
exist and point to where they are built if they aren't found.

finally, remove some commented out lines that gave alternatives
for how to name the files and directories, and other windows.
they looked pretty specific and users can change the location
of the input and output files as they'd like.


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@5934 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 55ad2ad9fa69c2d087b5c92f4cf2c2fd84fc96b9