Clean separation between CESM multi-instance setup and DART setup. Better default cutoff in input.nml.
The hard part is defining the output frequency of the history file
used for the flux tower observations. There is some argument to
make the history files be a full day long - which means that has
to be passed along to the hist_nhtfrq DART namelist item so it knows
how much time to subtract from the current model time to build the
CLM .h1. filename.  Even if the file contains a full day - it will still
support a 6 hour assimilation interval. For that matter, it could contain
a month. The forward operator code finds the history file time closest
to the observation time. Since the history file output frequency cannot
be changed after startup, the initial configuration must be preserved
in the DART namelist. The CLM run-time setting is meaningless.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@6411 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: d698cb9e78ddcfd7908ef28cd41d1dc1a05ed87d