The input namelist now REQUIRES 4 pieces of information for each input variable.
state_variables = 'VTEC',  'KIND_VERTICAL_TEC',              'NA',      'NA',
                  'TN',    'KIND_TEMPERATURE',               '0.0',     '6000.0',
                  'TN_NM', 'KIND_TEMPERATURE',               '0.0',     '6000.0',
                  'O1',    'KIND_ATOMIC_OXYGEN_MIXING_RATIO','0.00001', '0.98888',
                  'O1_NM', 'KIND_ATOMIC_OXYGEN_MIXING_RATIO','0.00001', '0.98888',
                  'NE',    'KIND_ELECTRON_DENSITY',          '1000.0',  'NA',

These limits are then imposed when the variable is updated in the TIEGCM restart file.
The values written to the DART netCDF files are NOT IMPOSED. If DART creates a 
temperature below absolute zero, the DART netCDF files will reflect the negative
temperature.

The get_val() routine is still unfinished. I believe it is supposed to
do the vertical interpolation for a given lat/lon. There is some extra
calculation for the vTEC observations, I think the whole routine could be
clearer. Also ... since ZG is local, it should be used - or do we use the
ensemble mean ZG?



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/tiegcm@6705 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 627391f906215f8a90a6b76295c2cf4be3928763