fix a problem with multiple nests and the option to increase obs errors near the domain boundaries. the index computations were using the innermost nest that contained the obs, but doing index computations relative to the outer domain. the result could be very large obs errors for obs near the edges of the nest. the resulting assimilation would then basically ignore the obs because the large error made it have no impact.
this was fixed by changing the get_domain_info() routine
to allow the caller to specify an optional domain number 
to start searching on.  without the argument the original
functionality remains unchanged.  the preprocess tool now
specifies domain 1 when looking for obs near the boundaries.

this preludes changing obs errors near inner nest
boundaries, which we have discussed and agreed is
not something we want to support at this time.

Reference jira DART-17.


git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@6877 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: a5d148f59b1bede17855c1e87696517485d5fbc6