add in code to support interpolation of locations outside the mpas grid but very close to the bottom or top levels.  allow user to say how close is close enough with a namelist item. leave a placeholder for real extrapolation, but for now just evaluate the locations as if they are on level 1 or N.
this satisfies a functionality soyoung wanted when she needed
to evaluate cell edges which were higher or lower than their
corresponding cell centers. 

also support an rbf width of 0, meaning only use the vertices 
in the current cell to compute the value.  

both of these functions may be most useful for debugging, but
still seem of value.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@6885 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 9732b0ee8e7b27be084017f794422b99484974ff