Converting the ZG variable directly to meters upon read. This fixed a problem that manifest itself when location_mod:horiz_dist_only = .FALSE. and then only with variables that used the 'ilev' dimension.
I added a matlab routine to explore the differences between two netCDF files 
for a set of variables. The intent is to run it with a known working Posterior_Diag.nc
and a candidate Posterior_Diag.nc.  Since the other variables in the netCDF files may differ,
this is not an obvious case for Compare_netCDF_files.m



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/tiegcm@7029 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 767e007e8752dffd001cda32232293bf9a2ab60b