the written documentation for the details of the POP grid is inconsistent with the current POP code.  the KMT array is read from a grid file which stores the lowest valid level in each column (0 = land).  an offset KMU array is computed from that for the velocity points.  the original code was using an offset in the wrong direction when trying to index the 4 T points that enclosed each U point.  it also did not handle the wrap-around in longitude correctly, leaving those KMU values undefined.
the impacts were:  possibly failed forward operators for U and V currents
if they were along the 'seam' in the shifted-pole grid; and current values 
in the ocean along the 'seam' not being impacted by observations when they should
have been.  the seam in the shifted grid runs along south america, up the
atlantic ocean, and into greenland.  the impacts are observable, but minimal.

the fix uses the right 4 KMT points to compute each KMU; it handles the
wraparound in longitude correctly; and it warns if the pole is unshifted
since we believe that the model itself might not correctly support a grid
with a singular point at the north pole.  this warning can be commented out
since the assimilation will work correctly with an unshifted grid; it is
more of a concern that the model advance will not have correctly computed
values at the pole.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@7223 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: f3362babd7359b391479e1c994ae6b503331afe3