The matlab plotting scripts annotate the fact that trusted observations were used. The number of observations assimilated remains unchanged, but the values plotted reflect the fact that outlier observations were used by obs_diag.
obs_diag now puts an attribute on each of the trusted variables to
identify which ones are trusted or not.

obs_diag also counts up the number of pathological observations that
have a DART QC of 7 but a missing posterior mean (would be a DART QC 4).
The count for the N_DARTqc_4 and N_DARTqc_7 are more accurate now than
previously. The metrics (rmse, bias, spread, etc) are unchanged.

This feature only affects the calculation of the metrics for TRUSTED observations.
If the observation is TRUSTED, then even the outlier obs are used to 
calculate the metrics. For those few with missing posterior means, every now
and again there are CRAZY posterior metrics because a MISSING posterior mean
is leaking into the calculation.  To compensate for this effect, there are
two lines of code that can be enabled to force the QC to be a 4. Those obs
cannot be used in the calculation of TRUSTED observations.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@7605 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: a02782495f4182bdeffeb370d8ef8497d223f40b