Added the observation converter to supply O/N2 ratio observations from SSUSI.
The original netCDF files must be preprocessed to add some netCDF variable
attributes to allow recognition of the (undocumented) _FillValue of NaN for
the ON2_UNCERTAINTY variable used to supply the observation error variance.
shell_scripts/netcdf_manip.csh does just that.

Added types to be able to distinguish VTEC and 'extended' VTEC and a stub routine
in the obs_def_upper_atm_mod.f90 to be used to replace the 
tiegcm/model_mod:create_vtec() routine.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/tiegcm@7651 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: d0918e134a6eb89180bd6839d106aaf055ac55a5