This has consistency between run_perfect_model_obs and run_filter Both of these scripts have improved comments and forcibly set the 'parent' tiegcm.nml and input.nml namelists in the run directory to have the values required by the advance_model.csh script.
The estimate_f107 option in the model_mod is set to cause a fatal
error if enabled. There are comments on how to enable it safely
under the assumption you know what you are doing. f107 now has
the 'local noon' location setting based on the seconds in the 
current tiegcm time.

advance_model.csh only changes the dynamic portion of the tiegcm.nml



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/tiegcm@7684 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 03e3b5cf1350599d9117dfb2ca573055176460cd