These are trivial changes. The tiegcm model has vertical coordinates of a magnitude that require more precision in a print statement in location_mod.f90
I added a comment in the obs_utilities_mod.f90 that the units of the
variable for observation error should be in 'units' of standard deviation.

The tiegcm quickbuild.csh removed the default namelists, which defeats the
whole point of making them. For consistency with the other models, they should
not be removed.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@7767 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: f04f9530c1ccd4ecfa86cb04e1287158a9d4fa18