This represents a big change to the TIEGCM model_mod. The composition of the DART state vector can now be specified at run time through the model_nml:variables construct. This has been implemented for more than a year on the tiegcm branch and is being merged onto the trunk. This change requires that static_init_model() have access to a tiegcm_restart_p.nc and tiegcm_s.nc file as well as a tiegcm.nml.
Since the tiegcm branch cannot be directly merged onto the trunk,
these changes are duplicates of those made to the branch. The revision
history only exists on the tiegcm branch unfortunately.

I am backing out a change made to location_mod.f90, that was not necessary.
The print statement has enough precision to handle up to 999 km, TIEGCM
never gets quite that high.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@7775 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: a55a5f04a1b8aaccd954f826ee77a445b260364c