The major upgrades and changes include the following. More detailed comments are interspersed in the log since late 2013. 1) Addition of WACCM to the list of CAM variants which can be used in assimilations.    This required the addition of a new vertical coordinate; scale height ('log_invP'),    which has a small but widespread footprint. 2) Addition of CAM-SE to the list of CAM variants.    This required new interpolation routines and grid file management to handle    the logically non-rectangular grid (cubed sphere).    This consists mostly of a large new section of the module, but some scattered    changes as well. 3) Fixing the misinterpretation of DART KINDs vs TYPEs.    This was fixed independently on the trunk. 4) Code review leading to some structural and extensive cosmetic changes to bring    it into line with DART practices. 5) Rewriting of get_close_obs to handle the DART QC status better and    improve the clarity and logic, as well as accomodate items 1)-4). 6) Implementation of an improved damping algorithm for the top layers of the state,    where CAM applies enhanced diffusion and the assimilation should be weaker.    This algorithm has the right units, works for vert_coord = 'pressure' and    'log_invP', and allows more of the state to feel the observations. 7) Inclusion of scripting which will enable running multiple assimilation cycles    in a single job, which eliminates much of the waiting in the queue, and saves    ~30% in computing time by running st_archive.sh offline as a single processor job.    This is accomplished through the creation of scripts ${CASE}.submit_cycles and    ${CASE}.run_cycles by the new CESM1_2_1_setup_advanced.  CESM1_2_1_setup_hybrid    now exists as the streamlined, 'beginners' script, which trades some of the    capabilities in CESM1_2_1_setup_advanced in favor of ease of use.
The CAM-FV version from branches/cam was tested against the trunk.
When the improvement in the choice of PS location in convert_vert is taken
into account, the results are identical.
No direct comparisons against the trunk are possible for WACCM or CAM-SE.
Those were compared qualitatively against the branches' CAM-FV assimilations.

A notable limitation discovered during testing is that filter can fail
for task counts > ~4200, due to an MPI imbalance.  Filter completes all of
its work, but never returns from MPI_finalize.  Work-arounds are available,
but are not included in the trunk.
At this time the scripts may only work for setting up the CAM5 version of WACCM.
The CAM4 version forecasts fail, probably due to input/forcing file incompatibility.
This is at the top of the 'fix-it' list.

These scripts do not support variable-resolution grids.
If you need them, they are available in branches/cam.



git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/trunk@7853 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 36c5e36ba8028a4b29fa9ff7bef99d0b5e19a806